+++ src/helpers.php
@@ -1,5 +1,7 @@
 <?php
 
+use TomatoPHP\FilamentMeta\Models\Meta;
+
 if (! function_exists('meta')) {
     function meta(
         string $key,
@@ -11,10 +13,10 @@ if (! function_exists('meta')) {
     ): mixed {
         if ($value !== null) {
             if ($value === 'null') {
-                return \TomatoPHP\FilamentMeta\Models\Meta::query()->updateOrCreate(['key' => $key], ['value' => null, 'key_value' => null]);
+                return Meta::query()->updateOrCreate(['key' => $key], ['value' => null, 'key_value' => null]);
             } else {
                 if ($type === 'key-value') {
-                    return \TomatoPHP\FilamentMeta\Models\Meta::query()->updateOrCreate(['key' => $key], [
+                    return Meta::query()->updateOrCreate(['key' => $key], [
                         'value' => null,
                         'key_value' => $value,
                         'type' => $type,
@@ -23,7 +25,7 @@ if (! function_exists('meta')) {
                         'response' => $response,
                     ]);
                 } else {
-                    return \TomatoPHP\FilamentMeta\Models\Meta::query()->updateOrCreate(['key' => $key], [
+                    return Meta::query()->updateOrCreate(['key' => $key], [
                         'value' => $value,
                         'type' => $type,
                         'date' => $date ?? now()->toDateString(),
@@ -33,7 +35,7 @@ if (! function_exists('meta')) {
                 }
             }
         } else {
-            $meta = \TomatoPHP\FilamentMeta\Models\Meta::query()->where('key', $key)->first();
+            $meta = Meta::query()->where('key', $key)->first();
             if ($meta) {
                 if ($type === 'key-value') {
                     return $meta->key_value;
@@ -43,7 +45,7 @@ if (! function_exists('meta')) {
 
             } else {
                 if (config('filament-meta.create')) {
-                    return \TomatoPHP\FilamentMeta\Models\Meta::query()->updateOrCreate(['key' => $key], [
+                    return Meta::query()->updateOrCreate(['key' => $key], [
                         'value' => null,
                         'type' => $type,
                         'date' => $date ?? now()->toDateString(),
@@ -53,5 +55,7 @@ if (! function_exists('meta')) {
                 }
             }
         }
+
+        return null;
     }
 }