+++ tests/src/TestCase.php
@@ -20,7 +20,7 @@ use Orchestra\Testbench\Concerns\WithWorkbench;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use RyanChandler\BladeCaptureDirective\BladeCaptureDirectiveServiceProvider;
 use TomatoPHP\FilamentMeta\FilamentMetaServiceProvider;
-use TomatoPHP\FilamentUsers\Tests\Models\User;
+use TomatoPHP\FilamentMeta\Tests\Models\User;
 
 #[WithEnv('DB_CONNECTION', 'testing')]
 abstract class TestCase extends BaseTestCase